注释调整

+++ services/distributeddataservice/service/data_share/data_share_service_impl.cpp
@@ -682,8 +682,8 @@ int32_t DataShareServiceImpl::OnAppExit(pid_t uid, pid_t pid, uint32_t tokenId, 
 {
     ZLOGI("AppExit uid=%{public}d, pid=%{public}d, tokenId=0x%{public}x, bundleName=%{public}s",
         uid, pid, tokenId, bundleName.c_str());
-    RdbSubscriberManager::GetInstance().Delete(tokenId,(uint32_t)pid);
-    PublishedDataSubscriberManager::GetInstance().Delete(tokenId,(uint32_t)pid);
+    RdbSubscriberManager::GetInstance().Delete(tokenId, pid);
+    PublishedDataSubscriberManager::GetInstance().Delete(tokenId, pid);
     return E_OK;
 }
